+++ frontend/st20/st20decoder.h
@@ -19,7 +19,7 @@
 
 #ifndef ST20DECODER
 #define ST20DECODER
-
+#include <cstddef>
 class Prog;
 class NJMCDecoder;
 struct DecodeResult;
@@ -35,13 +35,13 @@ public:
          * Decodes the machine instruction at pc and returns an RTL instance for
          * the instruction.
          */
-    virtual DecodeResult& decodeInstruction (ADDRESS pc, int delta);
+    virtual DecodeResult& decodeInstruction (ADDRESS pc, ptrdiff_t delta);
 
     /*
          * Disassembles the machine instruction at pc and returns the number of
          * bytes disassembled. Assembler output goes to global _assembly
          */
-    virtual int decodeAssemblyInstruction (ADDRESS pc, int delta);
+    virtual int decodeAssemblyInstruction (ADDRESS pc, ptrdiff_t delta);
 
 
 private:
