+++ include/cluster.h
@@ -11,7 +11,7 @@
  * \file       cluster.h
  * OVERVIEW:   Definition of the classes that describe a Cluster, a grouping
  *           of functions irrespective of relationship.  For example, the
- *           Object Oriented Programming concept of a Class is a Cluster. 
+ *           Object Oriented Programming concept of a Class is a Cluster.
  *           Clusters can contain other Clusters to form a tree.
  *============================================================================*/
 
@@ -46,11 +46,11 @@ protected:
 public:
                     Cluster() { }
                     Cluster(const std::string &_name) : name(_name){ }
-virtual                ~Cluster() {}
-        const        char *getName() { return name.c_str(); }
+virtual             ~Cluster() {}
+        const char *getName() { return name.c_str(); }
         void        setName(const std::string &nam) { name = nam; }
         size_t      getNumChildren() { return children.size(); }
-        Cluster *   getChild(int n) { return children[n]; }
+        Cluster *   getChild(size_t n) { return children[n]; }
         void        addChild(Cluster *n);
         void        removeChild(Cluster *n);
         Cluster *   getParent() { return parent; }
@@ -59,14 +59,12 @@ virtual                ~Cluster() {}
         void        openStreams(const char *ext);
         void        closeStreams();
         std::ofstream &getStream() { return out; }
-        const char *makeDirs();
-        const char *getOutPath(const char *ext);
+        const char* makeDirs();
+        const char* getOutPath(const char *ext);
         Cluster *   find(const std::string &nam);
 virtual bool        isAggregate() { return false; }
-
         void        printTree(std::ostream &out);
 protected:
-
     friend class XMLProgParser;
 };
 
@@ -84,7 +82,7 @@ protected:
 public:
     Class(const char *name) : Cluster(name) { type = new CompoundType(); }
 
-    // A Class tends to be aggregated into the parent Module, 
+    // A Class tends to be aggregated into the parent Module,
     // this isn't the case with Java, but hey, we're not doing that yet.
     virtual bool isAggregate() { return true; }
 };
